Modification PR#1177
