+++ open_spiel/python/CMakeLists.txt
@@ -297,6 +297,7 @@ if (OPEN_SPIEL_ENABLE_PYTORCH)
       pytorch/policy_gradient_pytorch_test.py
       pytorch/ppo_pytorch_test.py
       pytorch/neurd_pytorch_test.py
+      mfg/algorithms/mfg_proximal_policy_optimization_pytorch_test.py
   )
 endif()
